tuto4

+++ tutorial/snippets/views/fbv.py
@@ -1,21 +1,23 @@
+from django.http import HttpResponse
+from rest_framework import status
+from rest_framework.decorators import api_view
+from rest_framework.response import Response
+
 from snippets.models import Snippet
 from snippets.serializers import SnippetSerializer
-from django.http import Http404
-from rest_framework.views import APIView
-from rest_framework.response import Response
-from rest_framework import status
 
 
-class SnippetList(APIView):
+@api_view(['GET', 'POST'])
+def snippet_list(request, format=None):
     """
     코드 조각을 모두 보여주거나 새 코드 조각을 만듭니다.
     """
-    def get(self, request, format=None):
+    if request.method == 'GET':
         snippets = Snippet.objects.all()
         serializer = SnippetSerializer(snippets, many=True)
         return Response(serializer.data)
 
-    def post(self, request, format=None):
+    elif request.method == 'POST':
         serializer = SnippetSerializer(data=request.data)
         if serializer.is_valid():
             serializer.save()
@@ -23,30 +25,31 @@ class SnippetList(APIView):
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
 
-class SnippetDetail(APIView):
+@api_view(['GET', 'PUT', 'PATCH', 'DELETE'])
+def snippet_detail(request, pk, format=None):
     """
     코드 조각 조회, 업데이트, 삭제
     """
-    def get_object(self, pk):
-        try:
-            return Snippet.objects.get(pk=pk)
-        except Snippet.DoesNotExist:
-            raise Http404
+    try:
+        snippet = Snippet.objects.get(pk=pk)
+    except Snippet.DoesNotExist:
+        return HttpResponse(status=status.HTTP_404_NOT_FOUND)
 
-    def get(self, request, pk, format=None):
-        snippet = self.get_object(pk)
+    if request.method == 'GET':
         serializer = SnippetSerializer(snippet)
         return Response(serializer.data)
 
-    def put(self, request, pk, format=None):
-        snippet = self.get_object(pk)
-        serializer = SnippetSerializer(snippet, data=request.data)
+    elif request.method == 'PUT' or request.method == 'PATCH':
+        serializer = SnippetSerializer(
+            snippet,
+            data=request.data,
+            partial=True if request.method == 'PATCH' else False
+        )
         if serializer.is_valid():
             serializer.save()
             return Response(serializer.data)
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
-    def delete(self, request, pk, format=None):
-        snippet = self.get_object(pk)
+    elif request.method == 'DELETE':
         snippet.delete()
-        return Response(status=status.HTTP_204_NO_CONTENT)
+        return HttpResponse(status=status.HTTP_204_NO_CONTENT)
